Final changeset plus User controller and service tests

+++ src/main/java/bg/codeacademy/spring/gossiptalks/dto/PageDto.java
@@ -12,6 +12,6 @@ import java.util.List;
 public class PageDto
 {
   private Integer         numberOfElemets;
-  private Integer         totalElements;
+  private Long            totalElements;
   private List<GossipDto> content;
 }
